Replace all bitfields which were only declared with "(un)signed" with their actual types. Let me know if I missed any. It would also be a good idea to test this commit in both x64 and x86.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@6232 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/AudioCommon/Src/AudioCommon.h
@@ -36,21 +36,21 @@ union UDSPControl
 	u16 Hex;
 	struct
 	{
-		unsigned DSPReset       : 1; // Write 1 to reset and waits for 0
-		unsigned DSPAssertInt   : 1;
-		unsigned DSPHalt        : 1;
+		u16 DSPReset       : 1; // Write 1 to reset and waits for 0
+		u16 DSPAssertInt   : 1;
+		u16 DSPHalt        : 1;
 
-		unsigned AI             : 1;
-		unsigned AI_mask        : 1;
-		unsigned ARAM           : 1;
-		unsigned ARAM_mask      : 1;
-		unsigned DSP            : 1;
-		unsigned DSP_mask       : 1;
+		u16 AI             : 1;
+		u16 AI_mask        : 1;
+		u16 ARAM           : 1;
+		u16 ARAM_mask      : 1;
+		u16 DSP            : 1;
+		u16 DSP_mask       : 1;
 
-		unsigned ARAM_DMAState  : 1; // DSPGetDMAStatus() uses this flag
-		unsigned DSPInitCode    : 1;
-		unsigned DSPInit        : 1; // DSPInit() writes to this flag
-		unsigned pad            : 4;
+		u16 ARAM_DMAState  : 1; // DSPGetDMAStatus() uses this flag
+		u16 DSPInitCode    : 1;
+		u16 DSPInit        : 1; // DSPInit() writes to this flag
+		u16 pad            : 4;
 	};
 	UDSPControl(u16 _Hex = 0) : Hex(_Hex) {}
 };
