+++ Source/Core/Core/Src/HW/VideoInterface.h
@@ -114,9 +114,9 @@ union UVIVerticalTimingRegister
 	u16 Hex;
 	struct
 	{
-		unsigned EQU	:	 4; // Equalization pulse in half lines
-		unsigned ACV	:	10; // Active video in lines per field (seems always zero)
-		unsigned		:	 2;
+		u16 EQU	:	 4; // Equalization pulse in half lines
+		u16 ACV	:	10; // Active video in lines per field (seems always zero)
+		u16		:	 2;
 	};
 	UVIVerticalTimingRegister(u16 _hex) { Hex = _hex;}
 	UVIVerticalTimingRegister() { Hex = 0;}
@@ -127,14 +127,14 @@ union UVIDisplayControlRegister
 	u16 Hex;
 	struct
 	{
-		unsigned ENB	:	1; // Enables video timing generation and data request
-		unsigned RST	:	1; // Clears all data requests and puts VI into its idle state
-		unsigned NIN	:	1; // 0: Interlaced, 1: Non-Interlaced: top field drawn at field rate and bottom field is not displayed
-		unsigned DLR	:	1; // Selects 3D Display Mode
-		unsigned LE0	:	2; // Display Latch; 0: Off, 1: On for 1 field, 2: On for 2 fields, 3: Always on
-		unsigned LE1	:	2;
-		unsigned FMT	:	2; // 0: NTSC, 1: PAL, 2: MPAL, 3: Debug
-		unsigned		:	6;
+		u16 ENB	:	1; // Enables video timing generation and data request
+		u16 RST	:	1; // Clears all data requests and puts VI into its idle state
+		u16 NIN	:	1; // 0: Interlaced, 1: Non-Interlaced: top field drawn at field rate and bottom field is not displayed
+		u16 DLR	:	1; // Selects 3D Display Mode
+		u16 LE0	:	2; // Display Latch; 0: Off, 1: On for 1 field, 2: On for 2 fields, 3: Always on
+		u16 LE1	:	2;
+		u16 FMT	:	2; // 0: NTSC, 1: PAL, 2: MPAL, 3: Debug
+		u16		:	6;
 	};
 	UVIDisplayControlRegister(u16 _hex) { Hex = _hex;}
 	UVIDisplayControlRegister() { Hex = 0;}
@@ -146,12 +146,12 @@ union UVIHorizontalTiming0
 	struct { u16 Lo, Hi; };
 	struct  
 	{		
-		unsigned HLW		:	9; // Halfline Width (W*16 = Width (720))
-		unsigned			:	7;
-		unsigned HCE		:	7; // Horizontal Sync Start to Color Burst End
-		unsigned			:	1;
-		unsigned HCS		:	7; // Horizontal Sync Start to Color Burst Start
-		unsigned			:	1;
+		u32 HLW		:	9; // Halfline Width (W*16 = Width (720))
+		u32			:	7;
+		u32 HCE		:	7; // Horizontal Sync Start to Color Burst End
+		u32			:	1;
+		u32 HCS		:	7; // Horizontal Sync Start to Color Burst Start
+		u32			:	1;
 	};
 };
 
@@ -161,11 +161,11 @@ union UVIHorizontalTiming1
 	struct { u16 Lo, Hi; };
 	struct  
 	{		
-		unsigned HSY		:	 7; // Horizontal Sync Width
-		unsigned HBE640		:	 9; // Horizontal Sync Start to horizontal blank end
-		unsigned			:	 1;
-		unsigned HBS640		:	 9; // Half line to horizontal blanking start
-		unsigned			:	 6;
+		u32 HSY		:	 7; // Horizontal Sync Width
+		u32 HBE640	:	 9; // Horizontal Sync Start to horizontal blank end
+		u32			:	 1;
+		u32 HBS640	:	 9; // Half line to horizontal blanking start
+		u32			:	 6;
 	};
 };
 
@@ -176,10 +176,10 @@ union UVIVBlankTimingRegister
 	struct { u16 Lo, Hi; };
 	struct  
 	{		
-		unsigned PRB		:	10; // Pre-blanking in half lines
-		unsigned			:	 6;
-		unsigned PSB		:	10; // Post blanking in half lines
-		unsigned			:	 6;
+		u32 PRB		:	10; // Pre-blanking in half lines
+		u32			:	 6;
+		u32 PSB		:	10; // Post blanking in half lines
+		u32			:	 6;
 	};
 };
 
@@ -190,10 +190,10 @@ union UVIBurstBlankingRegister
 	struct { u16 Lo, Hi; };
 	struct  
 	{		
-		unsigned BS0		:	 5; // Field x start to burst blanking start in halflines
-		unsigned BE0		:	11; // Field x start to burst blanking end in halflines
-		unsigned BS2		:	 5; // Field x+2 start to burst blanking start in halflines
-		unsigned BE2		:	11; // Field x+2 start to burst blanking end in halflines
+		u32 BS0		:	 5; // Field x start to burst blanking start in halflines
+		u32 BE0		:	11; // Field x start to burst blanking end in halflines
+		u32 BS2		:	 5; // Field x+2 start to burst blanking start in halflines
+		u32 BE2		:	11; // Field x+2 start to burst blanking end in halflines
 	};
 };
 
@@ -204,11 +204,11 @@ union UVIFBInfoRegister
 	struct  
 	{
 		// TODO: mask out lower 9bits/align to 9bits???
-		unsigned FBB		:	24; // Base address of the framebuffer in external mem
+		u32 FBB		:	24; // Base address of the framebuffer in external mem
 		// POFF only seems to exist in the top reg. XOFF, unknown.
-		unsigned XOFF		:	 4; // Horizontal Offset of the left-most pixel within the first word of the fetched picture
-		unsigned POFF		:	 1; // Page offest: 1: fb address is (address>>5)
-		unsigned CLRPOFF	:	 3; // ? setting bit 31 clears POFF
+		u32 XOFF	:	 4; // Horizontal Offset of the left-most pixel within the first word of the fetched picture
+		u32 POFF	:	 1; // Page offest: 1: fb address is (address>>5)
+		u32 CLRPOFF	:	 3; // ? setting bit 31 clears POFF
 	};
 };
 
@@ -219,13 +219,13 @@ union UVIInterruptRegister
 	struct { u16 Lo, Hi; };
 	struct  
 	{		
-		unsigned HCT		:	11; // Horizontal Position
-		unsigned			:	 5;
-		unsigned VCT		:	11; // Vertical Position
-		unsigned			:	 1;
-		unsigned IR_MASK	:	 1; // Interrupt Mask Bit
-		unsigned			:	 2;
-		unsigned IR_INT		:	 1; // Interrupt Status (1=Active, 0=Clear)
+		u32 HCT		:	11; // Horizontal Position
+		u32			:	 5;
+		u32 VCT		:	11; // Vertical Position
+		u32			:	 1;
+		u32 IR_MASK	:	 1; // Interrupt Mask Bit
+		u32			:	 2;
+		u32 IR_INT	:	 1; // Interrupt Status (1=Active, 0=Clear)
 	};
 };
 
@@ -235,11 +235,11 @@ union UVILatchRegister
 	struct { u16 Lo, Hi; };
 	struct  
 	{		
-		unsigned HCT		:	11; // Horizontal Count
-		unsigned			:	 5;
-		unsigned VCT		:	11; // Vertical Count
-		unsigned			:	 4;
-		unsigned TRG		:	 1; // Trigger Flag
+		u32 HCT		:	11; // Horizontal Count
+		u32			:	 5;
+		u32 VCT		:	11; // Vertical Count
+		u32			:	 4;
+		u32 TRG		:	 1; // Trigger Flag
 	};
 };
 
@@ -248,8 +248,8 @@ union UVIHorizontalStepping
 	u16 Hex;
 	struct
 	{
-		unsigned FbSteps	:	 8;
-		unsigned FieldSteps	:	 8;
+		u16 FbSteps		:	 8;
+		u16 FieldSteps	:	 8;
 	};
 };
 
@@ -258,10 +258,10 @@ union UVIHorizontalScaling
 	u16 Hex;
 	struct
 	{
-		unsigned STP		:	9; // Horizontal stepping size (U1.8 Scaler Value) (0x160 Works for 320)
-		unsigned			:	3;
-		unsigned HS_EN		:	1; // Enable Horizontal Scaling
-		unsigned			:	3;
+		u16 STP		:	9; // Horizontal stepping size (U1.8 Scaler Value) (0x160 Works for 320)
+		u16			:	3;
+		u16 HS_EN	:	1; // Enable Horizontal Scaling
+		u16			:	3;
 	};
 	UVIHorizontalScaling(u16 _hex) { Hex = _hex;}
 	UVIHorizontalScaling() { Hex = 0;}
@@ -274,10 +274,10 @@ union UVIFilterCoefTable3
 	struct { u16 Lo, Hi; };
 	struct
 	{
-		unsigned Tap0		:	10;
-		unsigned Tap1		:	10;
-		unsigned Tap2		:	10;
-		unsigned			:	 2;
+		u32 Tap0	:	10;
+		u32 Tap1	:	10;
+		u32 Tap2	:	10;
+		u32			:	 2;
 	};
 };
 
@@ -288,10 +288,10 @@ union UVIFilterCoefTable4
 	struct { u16 Lo, Hi; };
 	struct
 	{
-		unsigned Tap0		:	 8;
-		unsigned Tap1		:	 8;
-		unsigned Tap2		:	 8;
-		unsigned Tap3		:	 8;
+		u32 Tap0	:	 8;
+		u32 Tap1	:	 8;
+		u32 Tap2	:	 8;
+		u32 Tap3	:	 8;
 	};
 };
 
@@ -308,10 +308,10 @@ union UVIBorderBlankRegister
 	struct { u16 Lo, Hi; };
 	struct
 	{
-		unsigned HBE656		:	10; // Border Horizontal Blank End
-		unsigned			:	11;
-		unsigned HBS656		:	10; // Border Horizontal Blank start
-		unsigned BRDR_EN	:	 1; // Border Enable
+		u32 HBE656	:	10; // Border Horizontal Blank End
+		u32			:	11;
+		u32 HBS656	:	10; // Border Horizontal Blank start
+		u32 BRDR_EN	:	 1; // Border Enable
 	};
 };
